Move parse expression logic to a separate file
